tools/doc-rag/requirements.txt
chromadb>=1.0
pymupdf>=1.24
python-docx>=1.0
fastmcp>=3.0
